cleanup store.js and seperate into smaller files
